+++ src/Builder/CardRequestBuilder.php
@@ -14,6 +14,13 @@
 
 namespace HiPay\PrestaShop\Builder;
 
+use HiPay\Fullservice\Enum\ThreeDSTwo\DeviceChannel;
+use HiPay\Fullservice\Enum\ThreeDSTwo\NameIndicator;
+use HiPay\Fullservice\Enum\ThreeDSTwo\ReorderIndicator;
+use HiPay\Fullservice\Gateway\Model\Request\ThreeDSTwo\AccountInfo;
+use HiPay\Fullservice\Gateway\Model\Request\ThreeDSTwo\BrowserInfo;
+use HiPay\Fullservice\Gateway\Model\Request\ThreeDSTwo\MerchantRiskStatement;
+use HiPay\Fullservice\Gateway\Model\Request\ThreeDSTwo\PreviousAuthInfo;
 use HiPay\Fullservice\Gateway\Request\Order\HostedPaymentPageRequest;
 use HiPay\Fullservice\Gateway\Request\Order\OrderRequest;
 use HiPay\PrestaShop\Settings\Entity\CardPaymentSettings;
@@ -59,7 +66,11 @@ class CardRequestBuilder extends AbstractPaymentRequestBuilder
      */
     public function buildRequest(bool $iframe = false)
     {
-        if (CardPaymentSettings::DISPLAY_MODE_HOSTED_FIELDS === $this->settings->cardPaymentSettings->displayMode && !isset($this->data['moto'])) {
+        $source = $this->data['source'] ?? '';
+        if (
+            (CardPaymentSettings::DISPLAY_MODE_HOSTED_FIELDS === $this->settings->cardPaymentSettings->displayMode && !isset($this->data['moto'])) ||
+            (CardPaymentSettings::DISPLAY_MODE_HOSTED_PAGE === $this->settings->cardPaymentSettings->displayMode && 'APPLE-PAY' === $source)
+        ) {
             return $this->buildOrderRequest();
         } else {
             return $this->buildHostedPaymentPageRequest($iframe);
@@ -75,6 +86,85 @@ class CardRequestBuilder extends AbstractPaymentRequestBuilder
         $request = new OrderRequest();
         $this->configureBaseFields($request);
         $request->paymentMethod->cardtoken = $this->data['token'] ?? '';
+        $request->custom_data = [
+            'multi_use' => $this->data['multi_use'] ?? null,
+            'one_click' => $this->data['one_click'] ?? null,
+        ];
+        if (isset($this->data['device_fingerprint'])) {
+            $request->device_fingerprint = $this->data['device_fingerprint'];
+        }
+        $browserInfo = new BrowserInfo();
+        $browserInfo->java_enabled = $this->data['browser_info']['java_enabled'] ?? null;
+        $browserInfo->javascript_enabled = $this->data['browser_info']['javascript_enabled'] ?? false;
+        $browserInfo->language = $this->data['browser_info']['language'] ?? null;
+        $browserInfo->color_depth = $this->data['browser_info']['color_depth'] ?? null;
+        $browserInfo->screen_height = $this->data['browser_info']['screen_height'] ?? null;
+        $browserInfo->screen_width = $this->data['browser_info']['screen_width'] ?? null;
+        $browserInfo->timezone = $this->data['browser_info']['timezone'] ?? null;
+        $browserInfo->http_user_agent = $this->data['browser_info']['http_user_agent'] ?? null;
+        $browserInfo->ipaddr = \Tools::getRemoteAddr();
+        $browserInfo->http_accept = $this->data['browser_info']['http_accept'] ?? null;
+        $request->browser_info = $browserInfo;
+        $request->device_channel = DeviceChannel::BROWSER;
+
+        $customer = new \Customer((int) $this->cart->id_customer);
+        if (!$customer->isGuest()) {
+            $merchantRiskStatement = new MerchantRiskStatement();
+            $merchantRiskStatement->reorder_indicator = $this->getOrderUniquenessInfo($customer->id, \Context::getContext()->cart->id) > 0 ? ReorderIndicator::REORDERED : ReorderIndicator::FIRST_TIME_ORDERED;
+            $request->merchant_risk_statement = $merchantRiskStatement;
+
+            $accountInfo = new AccountInfo();
+
+            $customerAccountInfo = new AccountInfo\Customer();
+            $customerAccountInfo->account_change = (int) date('Ymd', (int) strtotime($customer->date_upd));
+            $customerAccountInfo->opening_account_date = (int) date('Ymd', (int) strtotime($customer->date_add));
+            $customerAccountInfo->password_change = (int) date('Ymd', (int) strtotime($customer->last_passwd_gen));
+
+            $paymentAccountInfo = new AccountInfo\Payment();
+            if (isset($this->data['one_click']) && $this->data['one_click']) {
+                $customerCard = \HiPayPaymentsCustomerCard::getCustomerCardByCustomerIdToken($customer->id, $this->data['token']);
+                if ($customerCard->id) {
+                    $paymentAccountInfo->enrollment_date = (int) date('Ymd', (int) strtotime($customerCard->date_add));
+                }
+            }
+
+            $purchaseAccountInfo = new AccountInfo\Purchase();
+            $customerOrders = \Order::getCustomerOrders($customer->id);
+            if ($customerOrders) {
+                $sixMonthAgo = new \DateTime('6 months ago');
+                $sixMonthAgo = $sixMonthAgo->format('Y-m-d H:i:s');
+                $customerOrders = array_filter($customerOrders, function($order) use ($sixMonthAgo) {
+                    return $order['date_add'] >= $sixMonthAgo;
+                });
+            }
+            $purchaseAccountInfo->count = count($customerOrders);
+
+            $shippingAccountInfo = new AccountInfo\Shipping();
+            $deliveryAddressId = \Context::getContext()->cart->id_address_delivery;
+            if ($customerOrders) {
+                $customerOrders = array_filter($customerOrders, function($order) use($deliveryAddressId) {
+                    return (int) $order['id_address_delivery'] === (int) $deliveryAddressId;
+                });
+                if (!empty($customerOrders)) {
+                    end($customerOrders);
+                    $lastOrder = current($customerOrders);
+                    $shippingAccountInfo->shipping_used_date = (int) date('Ymd', (int) strtotime($lastOrder['date_add']));
+                }
+            }
+            $deliveryAddress = new \Address((int) $deliveryAddressId);
+            $customerFullName = strtoupper($customer->firstname.$customer->lastname);
+            $shippingFullName = strtoupper($deliveryAddress->firstname.$deliveryAddress->lastname);
+            $shippingAccountInfo->name_indicator = NameIndicator::DIFFERENT;
+            if ($customerFullName === '' || $customerFullName === $shippingFullName) {
+                $shippingAccountInfo->name_indicator = NameIndicator::IDENTICAL;
+            }
+
+            $accountInfo->customer = $customerAccountInfo;
+            $accountInfo->payment = $paymentAccountInfo;
+            $accountInfo->purchase = $purchaseAccountInfo;
+            $accountInfo->shipping = $shippingAccountInfo;
+            $request->account_info = $accountInfo;
+        }
 
         return $request;
     }
@@ -96,5 +186,37 @@ class CardRequestBuilder extends AbstractPaymentRequestBuilder
         }
 
         return $request;
+    }
+
+    /**
+     * @param int $idCustomer
+     * @param int $idCart
+     * @return int
+     */
+    private function getOrderUniquenessInfo(int $idCustomer, int $idCart): int
+    {
+        $sql = '
+            SELECT COUNT(*)
+            FROM (
+                SELECT o_hist.id_order
+                FROM '._DB_PREFIX_.'cart_product cp_ref
+                JOIN '._DB_PREFIX_.'cart c
+                    ON  c.id_cart     = cp_ref.id_cart
+                    AND c.id_customer = '.(int) $idCustomer.'
+                JOIN ' . _DB_PREFIX_.'orders o_hist
+                    ON o_hist.id_customer = '.(int) $idCustomer . '
+                JOIN '._DB_PREFIX_.'order_detail od_hist
+                    ON  od_hist.id_order             = o_hist.id_order
+                    AND od_hist.product_id           = cp_ref.id_product
+                    AND od_hist.product_attribute_id = cp_ref.id_product_attribute
+                    AND od_hist.product_quantity     = cp_ref.quantity
+                WHERE cp_ref.id_cart = '.(int) $idCart.'
+                GROUP BY o_hist.id_order
+                HAVING
+                    COUNT(*) = (SELECT COUNT(*) FROM '._DB_PREFIX_.'cart_product WHERE id_cart = '.(int) $idCart.')
+                    AND COUNT(*) = (SELECT COUNT(*) FROM '._DB_PREFIX_.'order_detail WHERE id_order = o_hist.id_order)
+            ) AS duplicates';
+
+        return (int) \Db::getInstance()->getValue($sql);
     }
 }