+++ views/js/ps-hosted-fields.js
@@ -304,7 +304,7 @@
             countryCode: PSHiPayData.cartDetails.countryCode,
             currencyCode: PSHiPayData.cartDetails.currencyCode,
             total: applePayTotal,
-            supportedNetworks: ['visa', 'masterCard']
+            supportedNetworks: ['visa', 'masterCard', 'cartesBancaires', 'maestro']
           };
 
           const applePayStyle = {
@@ -387,10 +387,8 @@
         }
         hipayPaymentsInstances[code].on('change', (event) => {
           if (!event.valid) {
-            document.querySelector('#payment-confirmation button').classList.add('disabled');
             document.querySelector('#payment-confirmation button').disabled = true;
           } else if (document.querySelector('input[name="conditions_to_approve[terms-and-conditions]"]').checked) {
-            document.querySelector('#payment-confirmation button').classList.remove('disabled');
             document.querySelector('#payment-confirmation button').disabled = false;
             document.getElementById(`js-hipay-payments-${code}-error-message`).innerHTML = '';
             document.getElementById(`js-hipay-payments-${code}-error-message`).style.display = 'none';