+++ CHANGELOG.md
@@ -4,6 +4,44 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.2] - 2026-03-23
+### Added
+
+- Added Cartes Bancaires as ApplePay supported networks
+- Added transaction reference in native table (Admin Order)
+- Added additional fields for DSP2
+
+### Changed
+
+- Changed lock type for notifications
+- Changed token storage strategy
+
+### Fixed
+
+- Fixed update from previous versions
+- Fixed Illicado with Hosted Fields
+- Fixed countries & currencies listings
+- Fixed Expired notifications for Alma
+- Fixed upgrade from previous versions
+- Fixed order request for ApplePay + HPP
+
+## [3.2.1] - 2026-03-16
+### Added
+
+- Added control on Oney phone numbers
+- Added logs on order validation
+
+### Changed
+
+- Updated CB logo
+
+### Fixed
+
+- Fixed status update (added priority)
+- Fixed UI on "Pay" button (PS1.7.6 / PS1.7.7)
+- Fixed possibility of orders duplicates
+- Fixed warning in logs file content
+
 ## [3.2.0] - 2026-01-31
 ### Added